+++ src/App.js
@@ -1,7 +1,7 @@
 import Footer from './components/Footer';
 import Header from './components/Header';
 import HomeScreen from './screens/HomeScreen';
-import ProductScreen from './screens/ProductScreen';
+import TodoScreen from './screens/TODOScreen';
 import {Container} from 'react-bootstrap'
 import {BrowserRouter as Router, Route, Routes} from 'react-router-dom'
 import LoginScreen from './screens/LoginScreen';
@@ -19,7 +19,7 @@ function App() {
               <Route path='/' element={<HomeScreen/>} exact/>
               <Route path='/login' element={<LoginScreen/>} />
               <Route path='/register' element={<RegisterScreens/>} />
-              <Route path='/product/:id' element={<ProductScreen/>} />
+              <Route path='/product/:id' element={<TodoScreen/>} />
               <Route path='/profile' element={<ProfileScreen/>} />
               <Route path='/ai-chat' element={<AiChatScreen/>}/>
             </Routes>
